cut_lors
